feat: add internal kubeconf gen to pub generators
