+++ pkg/clusters/addons/knative/knative.go
@@ -14,8 +14,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/kong/kubernetes-testing-framework/internal/utils"
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
+	"github.com/kong/kubernetes-testing-framework/pkg/utils/kubernetes/generators"
 )
 
 // -----------------------------------------------------------------------------
@@ -85,7 +85,7 @@ const (
 
 func deployKnative(ctx context.Context, cluster clusters.Cluster) error {
 	// generate a temporary kubeconfig since we use kubectl to deploy this addon
-	kubeconfig, err := utils.TempKubeconfig(cluster)
+	kubeconfig, err := generators.TempKubeconfig(cluster)
 	if err != nil {
 		return err
 	}
@@ -171,7 +171,7 @@ func deployKnative(ctx context.Context, cluster clusters.Cluster) error {
 
 func deleteKnative(ctx context.Context, cluster clusters.Cluster) error {
 	// generate a temporary kubeconfig since we use kubectl to cleanup this addon
-	kubeconfig, err := utils.TempKubeconfig(cluster)
+	kubeconfig, err := generators.TempKubeconfig(cluster)
 	if err != nil {
 		return err
 	}
