+++ pkg/clusters/addons/metallb/metallb.go
@@ -16,10 +16,10 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/kong/kubernetes-testing-framework/internal/utils"
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters/types/kind"
 	"github.com/kong/kubernetes-testing-framework/pkg/utils/docker"
+	"github.com/kong/kubernetes-testing-framework/pkg/utils/kubernetes/generators"
 	"github.com/kong/kubernetes-testing-framework/pkg/utils/networking"
 )
 
@@ -63,7 +63,7 @@ func (a *addon) Delete(ctx context.Context, cluster clusters.Cluster) error {
 	}
 
 	// generate a temporary kubeconfig since we're going to be using kubectl
-	kubeconfig, err := utils.TempKubeconfig(cluster)
+	kubeconfig, err := generators.TempKubeconfig(cluster)
 	if err != nil {
 		return err
 	}
@@ -198,7 +198,7 @@ address-pools:
 //       See: https://github.com/Kong/kubernetes-testing-framework/issues/25
 func metallbDeployHack(cluster clusters.Cluster) error {
 	// generate a temporary kubeconfig since we're going to be using kubectl
-	kubeconfig, err := utils.TempKubeconfig(cluster)
+	kubeconfig, err := generators.TempKubeconfig(cluster)
 	if err != nil {
 		return err
 	}
